+++ src/components/pages/HomePage/HomePage.tsx
@@ -1,7 +1,5 @@
 import { HomePageTemplate } from '@/components/templates/HomePageTemplate/HomePageTemplate';
 
-const HomePage = () => {
+export const HomePage = () => {
   return <HomePageTemplate />;
 };
-
-export default HomePage;
