+++ src/layouts/BottomNavLayout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
-import BottomNavbar from '@/components/organisms/BottomNavBar/BottomNavbar';
+import { BottomNavbar } from '@/components/organisms/BottomNavBar/BottomNavbar';
 
-const BottomNavLayout = () => {
+export const BottomNavLayout = () => {
   return (
     <div className="flex justify-center bg-gray-100">
       <div className="flex h-dvh w-dvw flex-col bg-white sm:w-[360px]">
@@ -15,5 +15,3 @@ const BottomNavLayout = () => {
     </div>
   );
 };
-
-export default BottomNavLayout;
